Applied linting across contracts except Hedera owned contracts (#28)

+++ contracts/IPair.sol
@@ -4,8 +4,7 @@ import "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
 import "./common/IBaseHTS.sol";
 import "./ILPToken.sol";
 
-abstract contract IPair { 
-
+abstract contract IPair {
     struct Pair {
         Token tokenA;
         Token tokenB;
@@ -13,13 +12,40 @@ abstract contract IPair {
 
     struct Token {
         address tokenAddress;
-        int tokenQty;
+        int256 tokenQty;
     }
 
     function getPair() external virtual returns (Pair memory);
-    function initialize(IBaseHTS _tokenService, ILPToken _lpTokenContract) public virtual;
-    function initializeContract(address fromAccount, address _tokenA, address _tokenB, int _tokenAQty, int _tokenBQty, int fee, address _treasury) external virtual;
-    function addLiquidity(address fromAccount, address _tokenA, address _tokenB, int _tokenAQty, int _tokenBQty) external virtual;
-    function removeLiquidity(address fromAccount, int _lpToken) external virtual;
-    function swapToken(address to, address _token, int _deltaQty) external virtual;
+
+    function initialize(IBaseHTS _tokenService, ILPToken _lpTokenContract)
+        public
+        virtual;
+
+    function initializeContract(
+        address fromAccount,
+        address _tokenA,
+        address _tokenB,
+        int256 _tokenAQty,
+        int256 _tokenBQty,
+        int256 fee,
+        address _treasury
+    ) external virtual;
+
+    function addLiquidity(
+        address fromAccount,
+        address _tokenA,
+        address _tokenB,
+        int256 _tokenAQty,
+        int256 _tokenBQty
+    ) external virtual;
+
+    function removeLiquidity(address fromAccount, int256 _lpToken)
+        external
+        virtual;
+
+    function swapToken(
+        address to,
+        address _token,
+        int256 _deltaQty
+    ) external virtual;
 }
